+++ src/main/java/com/example/models/DriverDto.java
@@ -4,22 +4,19 @@ import com.example.database.models.commons.DriverStatus;
 
 public class DriverDto {
 
-    private int id;
-
+    private Long id;
     private String driverLicense;
-
     private DriverStatus status;
-
     private UserDto user;
 
     public DriverDto() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
