+++ src/main/java/com/example/services/CargoService.java
@@ -7,11 +7,11 @@ public interface CargoService {
 
     CargoDto findById(Long cargoId);
     List<CargoDto> findAll();
-    CargoDto getCargoByDriverId(int driverId);
+    CargoDto getCargoByDriverId(Long driverId);
     CargoDto getCargoByTruckId(int truckId);
     boolean addCargo(CargoDto cargoDto);
     boolean updateCargo(CargoDto cargoDto);
-    boolean setAcceptStatus(Long cargoId, int driverId);
-    boolean setRefuseStatus(Long cargoId, int driverId);
-    boolean setDeliverStatus(Long cargoId, int driverId);
+    boolean setAcceptStatus(Long cargoId, Long driverId);
+    boolean setRefuseStatus(Long cargoId, Long driverId);
+    boolean setDeliverStatus(Long cargoId, Long driverId);
 }
